Fix MLX bfloat diagnostics conversion

+++ experiments/rag_vs_pra/run_powered_decomposition.py
@@ -312,7 +312,8 @@ class PersistentMLXBackend:
         targets = mx.array(answer_tokens, dtype=mx.int32)[:, None]
         selected_logits = mx.take_along_axis(answer_logits, targets, axis=-1).squeeze(-1)
         log_probabilities = selected_logits - mx.logsumexp(answer_logits, axis=-1)
-        first_logits = answer_logits[0]
+        log_probabilities = log_probabilities.astype(mx.float32)
+        first_logits = answer_logits[0].astype(mx.float32)
         mx.eval(log_probabilities, first_logits)
         values = np.asarray(log_probabilities, dtype=np.float64)
         first = np.asarray(first_logits, dtype="<f4")
